fix(timer-picker): address CodeRabbit findings on #55
- Move PresetSelection / resolvePreset / isCustomValid out of
  src/components/ into src/lib/timerPreset.ts. The hook
  useNewFocusWindow was importing domain helpers from a component
  module; per CLAUDE.md frontend rules, hooks/ should depend on lib/
  or api/, not on components/. TimerPresetPicker stays the UI piece
  and now just imports PresetSelection from the lib module.
- Reset PigDetail's timer-picker state when focus.id changes.
  Previously the same component instance could leak selection and
  custom-minutes across focuses when its focus prop swapped.

Test coverage:
- Move resolvePreset + isCustomValid unit tests to lib/timerPreset.test.ts.
- New PigDetail test renders a Harness that swaps the focus.id mid-flight
  and asserts the picker resets to the default selection.
